feat: Implement Phase 2 of Cycle Review - Frontend Structure
Implement all Phase 2 requirements from design document:
✅ 1. Create new route `/cycle-review` with React Router
✅ 2. Build CycleReviewPage component with navigation
✅ 3. Implement cycle selector dropdown for completed cycles
✅ 4. Create overview cards with comprehensive stats

Features implemented:
- Added React Router navigation between Planning and Cycle Review
- Created CycleReviewPage with cycle selection and data fetching
- Built OverviewCards component with 4 main metrics + velocity
- Added useCycleReviewData hook for API integration
- Structured PlanningPage as separate route from main App
- Added project and engineer preview cards
- Comprehensive error handling and loading states
- Auto-selects most recent completed cycle

UI includes:
- Clean navigation header with tab-based routing
- Cycle selector for completed cycles only
- Overview cards showing issues, points, PRs, team size
- Velocity calculation and preview of top projects/contributors
- Placeholder for export functionality (Phase 5)
- Tab navigation placeholder for detailed views (Phase 3)

Backend integration:
- Connects to existing /api/completed-cycles endpoint
- Integrates with /api/cycle-review/:cycleId endpoint
- Real-time data fetching with React Query caching
